requirements.txt
Django>=5.0,<5.1
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2.0
reportlab>=4.0.0